Correct fix for ftello64 etc.
ftell and fseek use "long" as an offset, which is probably OK on 64-bit unix platforms
where long is 64 bits, but will truncate offsets to 32 bits on, say, 32-bit Linux.

The warning about fseeko64 from an earlier PR was about how that is nonstandard.
fseeko, according to the FreeBSD manpage, conforms to: POSIX.1-2001, POSIX.1-2008, SUSv2